requirements.txt
django-crispy-forms==1.7.2
aniso8601==6.0.0
Django==2.2.4
django-mysql==3.2.0
djangorestframework==3.9.4
mysqlclient==1.3.14
pytz==2019.1
singledispatch==3.4.0.3
sqlparse==0.3.0